+++ types/metadataV10.go
@@ -22,9 +22,9 @@ import (
 	"hash"
 	"strings"
 
+	"github.com/Snowfork/go-substrate-rpc-client/blake2b"
 	"github.com/Snowfork/go-substrate-rpc-client/scale"
 	"github.com/Snowfork/go-substrate-rpc-client/xxhash"
-	"golang.org/x/crypto/blake2b"
 )
 
 // Modelled after packages/types/src/Metadata/v10/Metadata.ts
